Move remaining controllers into feature modules
