+++ pyproject.toml
@@ -43,7 +43,6 @@ zapzap = "zapzap.__main__:main"
 packages = [
   "zapzap",
   "zapzap.app",
-  "zapzap.controllers",
   "zapzap.core",
   "zapzap.core.config",
   "zapzap.core.diagnostics",
@@ -79,9 +78,13 @@ packages = [
   "zapzap.features.alerts",
   "zapzap.features.customizations",
   "zapzap.features.dictionaries",
+  "zapzap.features.donation",
   "zapzap.features.downloads",
+  "zapzap.features.downloads.ui",
   "zapzap.features.notifications",
+  "zapzap.features.onboarding",
   "zapzap.features.permissions",
+  "zapzap.features.shortcuts",
   "zapzap.features.startup",
   "zapzap.features.tray",
   "zapzap.assets",
