+++ zapzap/app/application.py
@@ -8,7 +8,7 @@ from PyQt6.QtGui import QDesktopServices
 
 from zapzap.app.single_application import SingleApplication
 from zapzap.app.startup_options import apply_startup_options, parse_startup_options
-from zapzap.controllers.onboarding_dialog import OnboardingDialog
+from zapzap.features.onboarding.onboarding_dialog import OnboardingDialog
 from zapzap.ui.main_window.client_side_rendering_controller import ClientSideRenderingController
 from zapzap.ui.main_window.main_window_controller import MainWindowController
 from zapzap.core.diagnostics import crash_handler
