+++ zapzap/features/donation/__init__.py
@@ -1,0 +1,1 @@
+"""Donation feature."""
