+++ zapzap/features/downloads/ui/__init__.py
@@ -1,0 +1,1 @@
+"""Download UI dialogs."""
