+++ zapzap/features/onboarding/__init__.py
@@ -1,0 +1,1 @@
+"""Onboarding feature."""
