+++ zapzap/features/shortcuts/__init__.py
@@ -1,0 +1,1 @@
+"""Shortcuts feature."""
