+++ zapzap/features/shortcuts/shortcuts_dialog.py
